+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Added overloads for `Sockets.findAvailableTcpPort`.
 - Added `Assert` utility class to be used to check preconditions.
+- Added `Bytes.padRightToBlockSize` to add padding to a byte array given a block size.
+- Added `Bytes.padPKCS7` to add PKCS#7 padding to a byte array.
 
 ---